LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-anchor-positioning-002-expected.txt


FAIL CSS Content Visibility: hidden never anchors assert_equals: #positioned should be anchored because it's not under a content-visibility ancestor expected 209 but got 8
